Created an AccessIndicator component
