+++ src/components/[hall]/GuildsByPlatform/components/GuildListItem/index.ts
@@ -1,0 +1,3 @@
+import GuildListItem from "./GuildListItem"
+
+export default GuildListItem
